Updated the adapter to refresh the recyclerview after adding a new word and result to the database.
Smooth scroll to final position of the word list, added LinearLayoutManager settings to the RecyclerView for reverse layout, and stack from end.

Commented out adding fake data for testing.

TODO: Need to convert the TRUE/FALSE tag for an image in place of the textview, green check and red x.